+++ src/main/java/com/prodapt/learningspring/repository/PostRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.prodapt.learningspring.entity.Post;
 
-public interface PostRepository extends CrudRepository<Post, Integer>{
+public interface PostRepository extends CrudRepository<Post, Long>{
 
 }
